+++ cli_tools/mympd-script/mympd-script.c
@@ -7,32 +7,36 @@
 #include "compile_time.h"
 
 #include "dist/sds/sds.h"
+#include "src/lib/filehandler.h"
 #include "src/lib/http_client.h"
 #include "src/lib/list.h"
 #include "src/lib/sds_extras.h"
 #include "cli_tools/log.h"
 
+#include <stdbool.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
 
 static void print_usage(char **argv) {
-    fprintf(stderr, "\nmyMPD script utility\n\n"
-        "Usage: %s <uri> <partition> <scriptname> [<arguments>]\n"
+    fprintf(stderr, "\nmyMPD script utility %s\n\n"
+        "Usage: %s [-k] <uri> <partition> <scriptname> [<arguments>]\n"
+        "  -k:           Disable certificate checking\n"
         "  <uri>:        myMPD listening uri, e.g. \"https://localhost\"\n"
         "  <partition>:  MPD partition, e.g. \"default\"\n"
         "  <scriptname>: script to execute, use \"-\" to read the script from stdin.\n"
         "  <arguments>:  optional space separated key=value pairs for script arguments.\n"
-        "\nFor further details look at https://jcorporation.github.io/myMPD/scripting/\n\n",
-        argv[0]);
+        "\n"
+        "For further details look at https://jcorporation.github.io/myMPD/050-scripting/mympd-script/\n\n",
+        MYMPD_VERSION, argv[0]);
 }
 
-static sds parse_arguments(sds post_data, char **argv, int argc) {
-    if (argc < 5) {
+static sds parse_arguments(sds post_data, char **argv, int argc, int arg_offset) {
+    if (argc < 5 + arg_offset) {
         return post_data;
     }
-    for (int i = 4; i < argc; i++) {
-        if (i > 4) {
+    for (int i = 4 + arg_offset; i < argc; i++) {
+        if (i > 4 + arg_offset) {
             post_data = sdscatlen(post_data, ",", 1);
         }
         int count = 0;
@@ -51,7 +55,16 @@ static sds parse_arguments(sds post_data, char **argv, int argc) {
 }
 
 int main(int argc, char **argv) {
-    if (argc < 4) {
+    bool cert_check = true;
+    int min_args = 4;
+    int arg_offset = 0;
+    if (argc > 1 && strcmp(argv[1], "-k") == 0) {
+        cert_check = false;
+        min_args++;
+        arg_offset++;
+    }
+
+    if (argc < min_args) {
         print_usage(argv);
         return EXIT_FAILURE;
     }
@@ -61,8 +74,8 @@ int main(int argc, char **argv) {
     sds uri = sdsempty();
     sds post_data = sdsempty();
 
-    if (strlen(argv[3]) == 1 && argv[3][0] == '-') {
-        uri = sdscatfmt(uri, "%s/script-api/%s", argv[1], argv[2]);
+    if (strlen(argv[3 + arg_offset]) == 1 && argv[3 + arg_offset][0] == '-') {
+        uri = sdscatfmt(uri, "%s/script-api/%s", argv[1 + arg_offset], argv[2 + arg_offset]);
         int c;
         sds script_data = sdsempty();
         while ((c = getchar()) != EOF) {
@@ -71,16 +84,16 @@ int main(int argc, char **argv) {
         post_data = sdscat(post_data, "{\"jsonrpc\":\"2.0\",\"id\":0,\"method\":\"INTERNAL_API_SCRIPT_POST_EXECUTE\",\"params\":{\"script\":");
         post_data = sds_catjson(post_data, script_data, sdslen(script_data));
         post_data = sdscat(post_data, ",\"arguments\":{");
-        post_data = parse_arguments(post_data, argv, argc);
+        post_data = parse_arguments(post_data, argv, argc, arg_offset);
         post_data = sdscat(post_data, "},\"event\":\"extern\"}");
         sdsfree(script_data);
     }
     else {
-        uri = sdscatfmt(uri, "%s/api/%s", argv[1], argv[2]);
+        uri = sdscatfmt(uri, "%s/api/%s", argv[1 + arg_offset], argv[2 + arg_offset]);
         post_data = sdscat(post_data, "{\"jsonrpc\":\"2.0\",\"id\":0,\"method\":\"MYMPD_API_SCRIPT_EXECUTE\",\"params\":{\"script\":");
-        post_data = sds_catjson(post_data, argv[3], strlen(argv[3]));
+        post_data = sds_catjson(post_data, argv[3 + arg_offset], strlen(argv[3 + arg_offset]));
         post_data = sdscat(post_data, ",\"arguments\":{");
-        post_data = parse_arguments(post_data, argv, argc);
+        post_data = parse_arguments(post_data, argv, argc, arg_offset);
         post_data = sdscat(post_data, "},\"event\":\"user\"}");
     }
 
@@ -88,8 +101,19 @@ int main(int argc, char **argv) {
         .method = "POST",
         .uri = uri,
         .extra_headers = "Content-type: application/json\r\n",
-        .post_data = post_data
+        .post_data = post_data,
+        .cert_check = cert_check
     };
+
+    if (cert_check == true) {
+        int nread;
+        sds ca_certs = sds_getfile(sdsempty(), CFG_MYMPD_CA_CERT_STORE, 1048576, false, true, &nread);
+        if (nread < 0) {
+            printf("ERROR: System certificate store not found.");
+            ca_certs = sdscat(ca_certs, "invalid");
+        }
+        request.ca_certs = ca_certs;
+    }
 
     struct mg_client_response_t response;
     http_client_response_init(&response);