+++ src/lib/http_client.c
@@ -203,6 +203,9 @@ static void http_client_ev_handler(struct mg_connection *nc, int ev, void *ev_da
             struct mg_tls_opts tls_opts = {
                 .name = host
             };
+            if (mg_client_request->cert_check == true) {
+                tls_opts.ca = mg_str(mg_client_request->ca_certs);
+            }
             mg_tls_init(nc, &tls_opts);
         }